refactor: merge utils/source.ts + views/widgets/displayLabel.ts → utils/sources.ts
Both modules did URL/path → label work; the duplication was the
exact "broken out by type" smell. Consolidate into one module:
srcKind() replaces _srcKind; labelFromUrl(src) ?? src replaces
_deriveLabel; toHttpsRepoUrl + labelFromManifest preserved as-is.

displayLabel.ts moves alongside the merge; source.ts is deleted.
