+++ app/src/coordinator.ts
@@ -24,7 +24,7 @@ import { buildFilePreviewPane, humanLanguageFor } from './views/panes/filePrevie
 import { buildCommitPane } from './views/panes/commitPane.js';
 import { buildStreetPane } from './views/panes/streetPane.js';
 import { sameDayCommitCount } from './utils/commitMetrics.js';
-import { labelFromManifest } from './views/widgets/displayLabel.js';
+import { labelFromManifest } from './utils/sources.js';
 import { LIVE_UPDATES } from './config/index.js';
 import { REBUILD_STATUS, LAST_REBUILD_ERROR, LAST_UPDATED_AT } from './store/liveStatus.js';
 import { DateSource, NodeKind } from './types';
